Added secure double JWT option
A secure double JWT saves the refresh_token in a secure httpOnly cookie that the client app cant access via JS, it returns the refresh_token_expiration date so that the client can handle the renewal, the cookie gets sent automatically with every request and is validated to provide an access_token, its recommended that the client app doesnt store the access_token anywhere so it can live safely in a memory state where an atacker cant access it, allows users to access via refresh- and access-tokens afterwards. Public contents are available with only a valid apikey.